ml-backend/requirements.txt
tensorflow==2.14.0
numpy==1.24.3
scikit-learn==1.3.0
flask==3.0.0
flask-cors==4.0.0
python-dotenv==1.0.0
pymongo==4.5.0
